testing conservation of angular momentum.
In the absence of external torques or forces, an articulated  model should not gain angular or linear momentum after an integration step.
This is a hard condition to achieve, since numerical integration introduces some errors,
but we expect that momentum changes should be small enough to be corrected by damping or control mechanisms.