display correlation between zero moment point and center of mass acceleration.
the objective function should now just minimize the center of mass horizontal acceleration
but the correction apply a constant of proportionality since the alpha is quite small in most cases.
the benefits of using center of mass acceleration is that is it generalized to un even terrains. at least I hope that's the case. but that for future tests.