adding gpu support to the Machine learning engine.
I'm using Vulkan for GPU acceleration in my project.
In performance comparisons between my radix sort implementations using CUDA and Vulkan,
I consistently observed that CUDA is more than twice as fast as Vulkan.

After extensive tuning and testing, I concluded that the main reason for this performance gap
is NVIDIA’s optimization strategy.
CUDA generates PTX(an intermediate representation), which NVIDIA appears to optimize very aggressively.
In contrast, SPIRV shaders used in Vulkan seems to be translated in a more direct manner,
with significantly fewer optimizations applied.

Essentially, it's like comparing a debug version of code to a release version and
expecting the debug version to be faster, something only possible if the debug version
uses a fundamentally more efficient algorithm.

This discrepancy becomes clear when compiling CUDA code with optimizations enabled;
under those conditions, the performance of Vulkan and CUDA can be comparable,
and in some cases, it's not obvious which one is faster.
Unfortunately, developers have limited control over this,
as it's up to NVIDIA to apply the same level of optimization to SPIRV as they do to PTX.

That said, Vulkan’s cross - platform support is a major strength.
Given the increasing diversity of hardware, the performance tradeoff is often acceptable.
Even with this overhead, Vulkan still significantly outperforms CPU - based implementations.
therefore the lost of performance only apply to nvidia hardware.