now I have all three back end working with general DeviceBuffers
last optimization is the indirect buffer minibatch loading.
The will load a input minibatch, using a minibatch size index array, that is loaded form host memory.
this index array implements the input shuffling, very import for training.
The hope is that loading the index array is faster than loading the minibatch for every step of training.