adding support for shared virtual memory support
course grained memory mapped is significant slower than not using it at all.
map / unmap has a significant over head cost.

The only alternative is just using fine grained virtual memory, which seem to be a but that is just kept mapped for ever. and the drive keep track of pci updates.
I seem the most risky too.

the cost is s till high but it seem to be the method wit the least overhead. so I keep it