anther fucking problem. Start make offset sizet_t instead of uint
using function in each class does no have the problem of vector size wrapping aorun the  2 billion elements.

after unifying the CUP and CPU, the buffer are now in bytes, which are 4 time larger that the number of floats.
I could make the buffer in int, but it is just a matter of time until that tow wrap around. therefre I better made that change now,.

the Nmist test reached about half billion float vectors, when multipath by the size of float, it is just under 2 billion bite offset, which is fine.
this is no the case for  the Reinforcement learning classes they have vector learning than half billion,  which definitely wrap around.