using pseudo subgroup for reduction.
if the number of item is less that the compute unit size, them reduction can be done without barriers, using shared memory.
amd wave front it 64, and nvidia 32, so I use 32
This trick is actually formalized in opencl with the cumbersome and difficult to use subgroup functionality.
therefore the soft implementation should be safe.