completed the teh GPU emutation
it is twice slower than the optimized cpu version.
the profile shows that the update is dominated by the two matrix multiplication which are time based in the emulation.
as I predicted this is twice slower, since the overhead for build the tile in cup cache is just too much to make up overhead.

this is because they are already in cache and they are read in coalesced stride, therefore tiling just double the bandwidth, and might even blow level 1 cache.

this is not the case for GPU, when the kernel control the shared memory, and the matrix are is main memory.

now next step is to revisit all the opencl kernels