make sure the cpu remain the must accurate and stable back end
the cpu optimized is the reference for all othe back ends.
Neural net operations are plagued of numerical errors, therefore it is very difficult to
check that on version was right or wrong, since training has the ability to adapt to lot of mistakes.
therefore, the way I found most efficient to debug this is by comparing with a correct version that I debug to make sure is numerically correct.