Update ndBrainAgentOffPolicyGradient_Trainer.cpp
testing learning the optimal sigma in training.
It seem to be slower to converge,  But in my opinion if I do no do that them the learning
is either too stochastic or too deterministic.
Of course I can aneling the constant sigma, but I want to try learning it form the q function. see how that works.