taken lower critic gradient is not better than the original paper.
It seem like is is slightly lower score by more stable.
I believe I read paper that did sated some alone that line.
Anyway stay with original method and do a final quarter of a million steps training.
I hope Hope is that the learning does nor degrades as it happened in the pass..