Update unicycle_sac.dnn
Just completed my first successful implementation of Soft Actor-Critic with entropy regularization that update the policy by calculation the entropy gradient.
IT is by far the best trainer I seen. :)

Next, I plan to experiment with autoregressive modeling of the entropy temperature.
However, I remain skeptical about the current literature on how entropy temperature is calculated.
My concern is that the temperature as a scalar coefficient, that is not part of the entropy, by the demonstrations use the Entropy whit is function of mean and variance..
In fact the use exoteric term like double meaning alpha,

Perhaps I’m simply missing something in the methodology, I probably need to dive deeper into the details.