removed support fo virtual memory mapped
out of resource may be a side effect of virtual memory mapped. making renamed buffers.
I am removing it since it does not really do anything it my king of hardware.
I might be good for apple arm, but apply does not support open cl cl, so there is
not reason for this functionality.