added hubber lost to the critic
it seem that the gain form the critic could be very larger causing the lease squared
error loss to be too extreme.
I am trying the Hubber loss with a max slope of 1.0
hopefully this is enough sufficient to proven exploding gradients after half million steps training. cruz my fingers.