making the thread pool adaptive
The original thread pool was designed with only a small number of cores in mind. It always activated all threads regardless of the workload size. On systems with many cores, this actually made things slower due to the overhead of synchronizing lots of small tasks.

The new thread system shifts more control to the client. Now, the client specifies how many threads the pool should use for a given task. On machines with only a few cores this change isn’t a big deal, but on high-core systems it makes a huge difference.

A nice side benefit is that CPU usage is now much lower, even for large scenes, since all threads are only dispatched for heavier tasks (which are the minority).

Another improvement is that on systems with a mix of big and small cores, the lighter threads tend to run on the low-ID cores more often. This gives the OS a clearer signal for managing CPU clock frequencies more efficiently