start adding a lightweight render library.
OpenGL is still my favorite rendering library, but with more platforms dropping support every month, it’s clear I need a replacement.

My first attempt was to move to Vulkan. After experimenting with it, though, I found it wasn’t worth the effort. It feels unnecessarily complex, and the platforms that don’t support OpenGL usually don’t support Vulkan either, or only do so through some form of emulation.

My second attempt was to look for a high-level rendering library, similar to what RenderWare offered back in the ’90s.
Unfortunately, that turned out to be a dead end due to endless dependency issues.

So, my conclusion is to encapsulate all rendering code into a separate library. Once I have it running with an OpenCL backend, I’ll add a Metal backend. My hope is that this will allow the engine to run smoothly on both Mac Minis and iPhones.