removed meshMatrix from scene node.
mesh matrix is used by Max, fbx, maya, and almost every the graphics package and file format but not by blender.
Blender has a delta transform matrix, but this is just an ill concept applied to the node matrix in global space.

In any case the mesh transform can be implemented in two ways:
-apply the transform to the mesh points.
-attach the mesh to a dummy node.

I am using method one.