add explicit path to all include.
cmake does no install the full include directory.
what people do is that the put the header is one subdirector and the code in another.
this what the install command only copy the file in the incode folder.

I prefer to have the include and the cpp in the same folder, this way the solution look simpler.
There are way around to this, but I think that make the path to include explicit at the project level
and the no adding relative path, is one of the simpler,

IT also allo for the end user to just opy the binary and use that