Update ndMeshImport.py
parsing a blender skeleton,  Blende skeletons are very strange.
The doc say the can be attach to a node, not I have found that quite difficult with the transition of editing mode. They seems to be designed to work best as root level objects.
anyway I now have to apply the transform.
hope this goes eassier.