replace shuffle with read from memory,
according to intel and some experts broadcast instructions have better throughput that shuffle.
so forcing the code to use shuffle with instruction lead to slower execution.
the problem is that broadcast cast is only available in avx, so using with intrinsic is define rabbit hole.
fortunately the compile does a very good job is with just asigning the single scalar.