start adding surrogate clip functions (wip)
so fat the variable sidestep performs a lot worse that the vanilla policy gradient, but is is stable.
It seem the surrogate clip function is quite important.

It also seem I was wrong assuming entropy is no important, because the object function is an entropy already.
but it seem that PPO do ten to close narrow that sigma and get stocked in a local minima.
It seem some reparametrized entropy could prevent this. anyway I continue.